ref(factory): Changed types used

+++ src/Factory/GetColor.cpp
@@ -14,11 +14,11 @@ namespace raytracer::math {
         }
         const auto color = shape["color"].as<>();
 
-        const long r = std::stol(color.substr(3, 4), nullptr, 16);
-        const long g = std::stol(color.substr(5, 6), nullptr, 16);
-        const long b = std::stol(color.substr(7, 8), nullptr, 16);
+        const double r = static_cast<double>(std::stol(color.substr(3, 4), nullptr, 16));
+        const double g = static_cast<double>(std::stol(color.substr(5, 6), nullptr, 16));
+        const double b = static_cast<double>(std::stol(color.substr(7, 8), nullptr, 16));
 
-        return Color(r / 255.0f, g / 255.0f, b / 255.0f);
+        return Color(r / 255.0, g / 255.0, b / 255.0);
     }
 
     Color
@@ -33,10 +33,10 @@ namespace raytracer::math {
         }
         const auto color = light["color"].as<>();
 
-        const long r = std::stol(color.substr(3, 4), nullptr, 16);
-        const long g = std::stol(color.substr(5, 6), nullptr, 16);
-        const long b = std::stol(color.substr(7, 8), nullptr, 16);
+        const double r = static_cast<double>(std::stol(color.substr(3, 4), nullptr, 16));
+        const double g = static_cast<double>(std::stol(color.substr(5, 6), nullptr, 16));
+        const double b = static_cast<double>(std::stol(color.substr(7, 8), nullptr, 16));
 
-        return Color(r / 255.0f, g / 255.0f, b / 255.0f);
+        return Color(r / 255.0, g / 255.0, b / 255.0);
     }
 }
